requirements.txt
telethon
python-dotenv
psycopg2-binary
dbt-postgres
ultralytics
fastapi
uvicorn
dagster
dagster-webserver
pydantic
ultralytics
